requirements.txt
numpy==2.2.4
PennyLane==0.41.0
Pillow==11.2.1
PyYAML==6.0.2
torch==2.6.0+cu126
torchmetrics==1.7.1
torchvision==0.21.0+cu126
tqdm==4.67.1
einops==0.6.1
matplotlib==3.7.2
opencv_python==4.8.0.74
